+++ packages/plugin-input-exercise/src/index.ts
@@ -2,6 +2,7 @@ import { StatefulPlugin, StateType } from '@edtr-io/core'
 
 import { InputExerciseEditor } from './editor'
 import { createIcon, faKeyboard } from '@edtr-io/editor-ui'
+import { createPluginTheme } from '@edtr-io/ui'
 
 export const wrongAnswerObject = StateType.object({
   value: StateType.string(''),
@@ -18,4 +19,18 @@ export const inputExercisePlugin: StatefulPlugin<typeof inputExerciseState> = {
   title: 'Eingabefeld',
   icon: createIcon(faKeyboard),
   description: 'Füge deiner Aufgabe ein Eingabefeld für die Lernenden hinzu.'
+}
+
+export const createInputExerciseTheme = createPluginTheme<InputExerciseTheme>(
+  theme => {
+    return {
+      borderColor: theme.renderer.primary.background,
+      borderStyle: '3px solid'
+    }
+  }
+)
+
+interface InputExerciseTheme {
+  borderColor: string
+  borderStyle: string
 }