+++ packages/ui-renderer/src/components/index.ts
@@ -1,4 +1,5 @@
 export * from './expandable-box'
 export * from './feedback'
+export * from './submit-button'
 export * from './fetch-dimensions'
 export * from './icon'