+++ scripts/changelog.ts
@@ -192,7 +192,8 @@ async function exec(): Promise<void> {
         '**plugin-rows**. Added more theming options',
         '**core**. Added `isEmpty(state: State, id: string)` to check if a plugin state is empty',
         '**core**. Added optional function `isEmpty: (state: StateDescriptorValueType<S>) => boolean` to plugin definition',
-        '**plugin-text**. Added visual latex editor'
+        '**plugin-text**. Added visual latex editor',
+        '**plugin-input-exercise: improved UI/UX and added theming options'
       ],
       fixed: ['**plugin-image**. Hide config-overlay in render mode'],
       internal: [